@@ -3,26 +3,27 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "firebase": "^9.9.0",
-    "html2canvas": "^1.0.0-rc.7",
+    "@reduxjs/toolkit": "^2.2.7",
+    "@testing-library/jest-dom": "^6.5.0",
+    "@testing-library/react": "^16.0.1",
+    "@testing-library/user-event": "^14.5.2",
+    "firebase": "^10.13.1",
+    "html2canvas": "^1.4.1",
     "immutability-helper": "^3.1.1",
-    "jspdf": "^2.3.1",
-    "node-sass": "^6.0.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-redux": "^8.0.2",
+    "jspdf": "^2.5.1",
+    "node-sass": "^9.0.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-redux": "^9.1.2",
     "react-redux-firebase": "^3.11.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
-    "redux": "^4.2.0",
+    "react-router-dom": "^6.26.2",
+    "react-scripts": "5.0.1",
+    "redux": "^5.0.1",
     "redux-devtools-extension": "^2.13.9",
-    "redux-firestore": "^1.0.0",
-    "redux-thunk": "^2.4.1",
-    "uuid": "^8.3.2",
-    "web-vitals": "^1.1.2"
+    "redux-firestore": "^2.0.1",
+    "redux-thunk": "^3.1.0",
+    "uuid": "^10.0.0",
+    "web-vitals": "^4.2.3"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -47,5 +48,8 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "ajv": "^7.2.4"
   }
 }
