@@ -1,121 +1,176 @@
-import React,{useState} from "react";
-import { NavLink } from 'react-router-dom';
+import React, { useState } from "react";
+import { NavLink } from "react-router-dom";
 import { withRouter } from "react-router-dom";
-import ResumePreview from './resumePreview'
-import {skinCodes, fieldCd} from './../../constants/typeCodes';
-import { connect } from 'react-redux'
-import {setEducation,updateEducation} from '../../redux/actions/educationActions'
+import ResumePreview from "./resumePreview";
+import { skinCodes, fieldCd } from "./../../constants/typeCodes";
+import { connect } from "react-redux";
+import {
+  setEducation,
+  updateEducation,
+} from "../../redux/actions/educationActions";
 // import {bindActionCreators} from 'redux';
-import { useHistory } from "react-router-dom";
+import { useNavigate } from "react-router-dom";
 
 function Education(props) {
+  console.log("Education");
 
-    console.log('Education');
+  let navigate = useNavigate();
+  const [education, addEducation] = useState(props.education);
 
-    let history = useHistory();
-    const [education,addEducation]= useState(props.education);
+  const onchange = (event) => {
+    var key = event.target.name;
+    var val = event.target.value;
+    addEducation({ ...education, [key]: val });
+  };
 
-    const onchange = (event) => {
-      var key =event.target.name;
-      var val =event.target.value;
-      addEducation({...education,[key]:val})
+  const getFieldData = (key) => {
+    if (education && education[key]) {
+      return education[key];
+    }
+    return "";
+  };
+
+  const onSubmit = async (e) => {
+    // console.log(this.state.educationSection);
+
+    if (props.education != null) {
+      props.updateEducation(education);
+    } else {
+      props.setEducation(education);
     }
 
-    const getFieldData=(key)=>{
-      if(education && education[key]){
-        return education[key]
-      }
-      return "";
-    }
+    navigate("/finalize");
+  };
 
-    const onSubmit = async(e) => {
-      // console.log(this.state.educationSection);
-
-      if(props.education!=null){
-          props.updateEducation(education);
-      }else{
-          props.setEducation(education);
-      }
-
-      history.push('/finalize')
-    }
-
-    return (
-      <div className="container med education" >
-        <div className="section funnel-section">
-          <div className="form-card">
-            <h2 className="form-heading center">Educational Section</h2>
-            <div className="form-section">
-              <div className="input-group"><label>College Name</label>
-                <div className="effect"><input type="text" name={fieldCd.SchoolName}
-                  onChange={onchange} value={getFieldData(fieldCd.SchoolName)} /><span></span>
-                </div>
-                <div className="error"></div>
+  return (
+    <div className="container med education">
+      <div className="section funnel-section">
+        <div className="form-card">
+          <h2 className="form-heading center">Educational Section</h2>
+          <div className="form-section">
+            <div className="input-group">
+              <label>College Name</label>
+              <div className="effect">
+                <input
+                  type="text"
+                  name={fieldCd.SchoolName}
+                  onChange={onchange}
+                  value={getFieldData(fieldCd.SchoolName)}
+                />
+                <span></span>
               </div>
+              <div className="error"></div>
+            </div>
 
-              <div className="input-group"><label>Degree</label>
-                <div className="effect"><input type="text" name={fieldCd.Degree}
-                  onChange={onchange} value={getFieldData(fieldCd.Degree)} /><span></span>
-                </div>
-                <div className="error"></div>
+            <div className="input-group">
+              <label>Degree</label>
+              <div className="effect">
+                <input
+                  type="text"
+                  name={fieldCd.Degree}
+                  onChange={onchange}
+                  value={getFieldData(fieldCd.Degree)}
+                />
+                <span></span>
               </div>
+              <div className="error"></div>
+            </div>
 
-              <div className="input-group"><label>CGPA</label>
-                <div className="effect"><input type="text" name={fieldCd.GraduationCGPA}
-                  onChange={onchange} value={getFieldData(fieldCd.GraduationCGPA)} /><span></span>
-                </div>
-                <div className="error"></div>
+            <div className="input-group">
+              <label>CGPA</label>
+              <div className="effect">
+                <input
+                  type="text"
+                  name={fieldCd.GraduationCGPA}
+                  onChange={onchange}
+                  value={getFieldData(fieldCd.GraduationCGPA)}
+                />
+                <span></span>
               </div>
+              <div className="error"></div>
+            </div>
 
-              <div className="input-group"><label>City/State</label>
-                <div className="effect"><input type="text"  name={fieldCd.City}
-                  onChange={onchange} value={getFieldData(fieldCd.City)} /><span></span>
-                </div>
-                <div className="error"></div>
+            <div className="input-group">
+              <label>City/State</label>
+              <div className="effect">
+                <input
+                  type="text"
+                  name={fieldCd.City}
+                  onChange={onchange}
+                  value={getFieldData(fieldCd.City)}
+                />
+                <span></span>
               </div>
+              <div className="error"></div>
+            </div>
 
-              <div className="input-group"><label>Graduation Month</label>
-                <div className="effect"><input type="text" name={fieldCd.GraduationDate}
-                  onChange={onchange} value={getFieldData(fieldCd.GraduationDate)} /><span></span>
-                </div>
-                <div className="error"></div>
+            <div className="input-group">
+              <label>Graduation Month</label>
+              <div className="effect">
+                <input
+                  type="text"
+                  name={fieldCd.GraduationDate}
+                  onChange={onchange}
+                  value={getFieldData(fieldCd.GraduationDate)}
+                />
+                <span></span>
               </div>
+              <div className="error"></div>
+            </div>
 
-              <div className="input-group"><label>Graduation Year</label>
-                <div className="effect"><input type="text"  name={fieldCd.GraduationYear}
-                  onChange={onchange} value={getFieldData(fieldCd.GraduationYear)} /><span></span>
-                </div>
-                <div className="error"></div>
+            <div className="input-group">
+              <label>Graduation Year</label>
+              <div className="effect">
+                <input
+                  type="text"
+                  name={fieldCd.GraduationYear}
+                  onChange={onchange}
+                  value={getFieldData(fieldCd.GraduationYear)}
+                />
+                <span></span>
               </div>
+              <div className="error"></div>
+            </div>
 
-              <div className="form-buttons">
-                <button className="btn hvr-float-shadow" type='button' onClick={onSubmit}>Next</button>
-                <NavLink to='/contact' className="center">Back</NavLink>
-              </div>
+            <div className="form-buttons">
+              <button
+                className="btn hvr-float-shadow"
+                type="button"
+                onClick={onSubmit}
+              >
+                Next
+              </button>
+              <NavLink to="/contact" className="center">
+                Back
+              </NavLink>
             </div>
           </div>
-          <div className="preview-card">
-            <ResumePreview contactSection={props.contact} educationSection={props.education} skinCd={props.document.skinCd}></ResumePreview>            
-          </div>
+        </div>
+        <div className="preview-card">
+          <ResumePreview
+            contactSection={props.contact}
+            educationSection={props.education}
+            skinCd={props.document.skinCd}
+          ></ResumePreview>
         </div>
       </div>
-    );
-  }
-  
-const mapStateToProps=(state)=>{
-    return{
-      document: state.document,
-      contact: state.contact,
-      education: state.education
-    }
+    </div>
+  );
 }
 
-const mapDispatchToProps=(dispatch)=>{
-    return{
-      setEducation: (education)=> dispatch(setEducation(education)),
-      updateEducation: (education)=> dispatch(updateEducation(education))
-    }
-}
+const mapStateToProps = (state) => {
+  return {
+    document: state.document,
+    contact: state.contact,
+    education: state.education,
+  };
+};
 
-export default connect(mapStateToProps,mapDispatchToProps)(Education);
+const mapDispatchToProps = (dispatch) => {
+  return {
+    setEducation: (education) => dispatch(setEducation(education)),
+    updateEducation: (education) => dispatch(updateEducation(education)),
+  };
+};
 
+export default connect(mapStateToProps, mapDispatchToProps)(Education);
